requirements.txt
anthropic
langgraph
langchain-anthropic
sentence-transformers
requests
beautifulsoup4
feedparser
playwright
apscheduler
streamlit
python-dotenv
pdfminer.six
python-docx
google-auth
google-auth-oauthlib
google-api-python-client
twilio
PyYAML
rich
loguru
numpy